move FloodgateBedrockPlatformChecker into commons
