+++ common/src/main/java/com/github/camotoy/geyserblockplatform/common/platformchecker/FloodgateBedrockPlatformChecker.java
@@ -1,14 +1,13 @@
-package com.github.camotoy.geyserblockplatform.bungeecord;
+package com.github.camotoy.geyserblockplatform.common.platformchecker;
 
 import com.github.camotoy.geyserblockplatform.common.device.DeviceOsFixer;
-import com.github.camotoy.geyserblockplatform.common.platformchecker.BedrockPlatformChecker;
 import org.geysermc.floodgate.api.FloodgateApi;
 import org.geysermc.floodgate.api.player.FloodgatePlayer;
 import org.geysermc.floodgate.util.DeviceOs;
 
 import java.util.UUID;
 
-public class FloodgateBungeeBedrockPlatformChecker implements BedrockPlatformChecker {
+public class FloodgateBedrockPlatformChecker implements BedrockPlatformChecker{
     @Override
     public DeviceOs getBedrockPlatform(UUID uuid) {
         FloodgatePlayer player = FloodgateApi.getInstance().getPlayer(uuid);
